# Allow Sum64String and (*Digest).WriteString to be inlined