--- a/xxhash_unsafe_test.go
+++ b/xxhash_unsafe_test.go
@@ -3,6 +3,8 @@
 package xxhash
 
 import (
+	"os/exec"
+	"sort"
 	"strings"
 	"testing"
 )
@@ -21,4 +23,39 @@ func TestStringAllocs(t *testing.T) {
 			sink = d.Sum64()
 		})
 	})
+}
+
+// This test is inspired by the Go runtime tests in https://golang.org/cl/57410.
+// It asserts that certain important functions may be inlined.
+func TestInlining(t *testing.T) {
+	funcs := map[string]struct{}{
+		"Sum64String":           {},
+		"(*Digest).WriteString": {},
+	}
+
+	// TODO: it would be better to use the go binary that is running
+	// 'go test' (if we are running under 'go test').
+	cmd := exec.Command("go", "test", "-gcflags=-m", "-run", "xxxx")
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		t.Log(string(out))
+		t.Fatal(err)
+	}
+
+	for _, line := range strings.Split(string(out), "\n") {
+		parts := strings.Split(line, ": can inline")
+		if len(parts) < 2 {
+			continue
+		}
+		delete(funcs, strings.TrimSpace(parts[1]))
+	}
+
+	var failed []string
+	for fn := range funcs {
+		failed = append(failed, fn)
+	}
+	sort.Strings(failed)
+	for _, fn := range failed {
+		t.Errorf("function %s not inlined", fn)
+	}
 }