@cespare 2020-11-20
Yes, that's because xxhashbench uses indirect calls. What I was getting at in https://github.com/cespare/xxhash/pull/42#pullrequestreview-535076183 and https://github.com/cespare/xxhash/issues/22#issuecomment-730782352 is that I don't think the indirect call performance matters that much.

I intend to delete xxhashbench (or at least rework it to use only direct calls). For now, I think we can focus on the new benchmarks I added in the xxhash package itself.