@cespare 2020-11-20
With help from @josharian I got a different way to write the conversions that allow both `Sum64String` and `(*Digest).WriteString` to be inlined. I updated the description and benchmarks.

I'm pretty happy with this for now so I think I'll merge.